requirements.txt
streamlit>=1.33
pandas>=2.1
numpy>=1.26
plotly>=5.18
openpyxl>=3.1
xlsxwriter>=3.1
reportlab>=4.0
python-dateutil>=2.9
requests>=2.32
charset-normalizer>=3.3
openai>=1.58
gspread>=6.1
google-auth>=2.37
